src/python/src/pyproject.toml
[tool.poetry]
name = "bb_proxy"
version = "0.1.0"
description = "Dep python proxy"
authors = ["Nevoliashka <41319097+Nevoliashka@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
mitmproxy = "^7.0.4"
click = "^8.0.4"
python-dotenv = "^0.19.2"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.32"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
